requirements.txt
boto3
click
pytest
pyyaml
